maven: Refactor testing integration tests into separate test projects
