+++ maven/bnd-testing-maven-plugin/src/it/parent/invoker.properties
@@ -1,0 +1,10 @@
+invoker.goals=--no-transfer-progress package
+
+# Build parent before other tests
+invoker.ordinal = 1000
+
+# Run mvn with --debug for debug logging
+#invoker.debug=true
+
+# Run mvn in debugging mode and wait for a debugger to attach
+#invoker.environmentVariables.MAVEN_DEBUG_OPTS=-agentlib:jdwp=transport=dt_socket,server=y,suspend=y,address=8000
