+++ maven/bnd-testing-maven-plugin/src/it/parent/pom.xml
@@ -14,13 +14,6 @@
 		</repository>
 	</distributionManagement>
 
-	<modules>
-		<module>test-no-resolve</module>
-		<module>test-with-resolve</module>
-		<module>test-with-resolve-from-dependencies</module>
-		<module>test-single-test</module>
-	</modules>
-
 	<dependencyManagement>
 		<dependencies>
 			<dependency>
@@ -100,7 +93,16 @@
 						</execution>
 					</executions>
 				</plugin>
-
+				<plugin>
+					<groupId>org.apache.maven.plugins</groupId>
+					<artifactId>maven-compiler-plugin</artifactId>
+					<version>3.10.1</version>
+				</plugin>
+				<plugin>
+					<groupId>org.apache.maven.plugins</groupId>
+					<artifactId>maven-resources-plugin</artifactId>
+					<version>3.2.0</version>
+				</plugin>
 				<!-- The following configuration is required because bnd-maven-plugin
 					generates the manifest to target/classes/META-INF/MANIFEST.MF but the normal
 					maven-jar-plugin does not use it. If the jar-plugin is patched to pick up
@@ -108,7 +110,7 @@
 				<plugin>
 					<groupId>org.apache.maven.plugins</groupId>
 					<artifactId>maven-jar-plugin</artifactId>
-					<version>3.2.0</version>
+					<version>3.2.2</version>
 					<configuration>
 						<archive>
 							<manifestFile>${project.build.outputDirectory}/META-INF/MANIFEST.MF</manifestFile>
