+++ maven/bnd-testing-maven-plugin/src/it/test-no-resolve/src/main/java/bnd/test/SimpleBndTest.java
