+++ maven/bnd-testing-maven-plugin/src/it/test-no-resolve/test.bndrun
