+++ maven/bnd-testing-maven-plugin/src/it/test-single-test/bnd.bnd
