+++ maven/bnd-testing-maven-plugin/src/it/test-single-test/src/main/java/bnd/test/FailingBndTest.java
