+++ maven/bnd-testing-maven-plugin/src/it/test-with-resolve-from-dependencies/bnd.bnd
