+++ maven/bnd-testing-maven-plugin/src/it/test-with-resolve-from-dependencies/invoker.properties
@@ -1,0 +1,7 @@
+invoker.goals=--no-transfer-progress integration-test
+
+# Run mvn with --debug for debug logging
+#invoker.debug=true
+
+# Run mvn in debugging mode and wait for a debugger to attach
+#invoker.environmentVariables.MAVEN_DEBUG_OPTS=-agentlib:jdwp=transport=dt_socket,server=y,suspend=y,address=8000
