+++ maven/bnd-testing-maven-plugin/src/it/test-with-resolve-from-dependencies/src/main/java/bnd/test/ResolvingBndTest.java
