+++ maven/bnd-testing-maven-plugin/src/it/test-with-resolve/src/main/java/bnd/test/ResolvingBndTest.java
