+++ maven/bnd-testing-maven-plugin/src/it/test-with-resolve/test.bndrun
